+++ src/common/mod.rs
@@ -3,6 +3,5 @@
 pub mod audio_source;
 pub mod batch_response;
 pub mod options;
-pub mod redacted_string;
 pub mod response;
 pub mod stream_response;
